started major refactoring
Data transformation / feature engineering has been moved to the end of the
notebook under a 'Model' section.
This will make creating several models easier.

Warning: a column in the dataset for which we want to make predictions seems
to have disappeared after the feature engineering phase.
We will need to investigate on that.
Anyway the number of columns is less than the number of columns in the training
dataset, this is a problem that needs to be solved.